Mostly and support for "System properties".
So properties set using -D  command line options will be consider first.
If editable and comming for system: if the value is change, the change
will be kept in preferences file BUT at run time the change will not be
accessible because the system property will still be taken FIRST.

A LogLevel property type has been added.